+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 # Changelog
+# Version 3.61.0 (2024-01-22)
+# Added
+* `ModelSlice.get_data_row_identifiers`
+  * Fetches all data row ids and global keys for the model slice
+  * NOTE Foundry model slices are note supported yet
+## Updated
+* Updated exports v1 deprecation date to April 30th, 2024
+* Remove `streamable` param from export_v2 methods
+
 # Version 3.60.0 (2024-01-17)
 ## Added
 * Get resource tags from a project